refactor: update exceptions handling flow
- add custom exception types PersistenceException, HostNotFoundException

- refactor exceptions handling to let them "bubble" upper to place where we can meaningfully handle them

- refactor handling "bubbled" exceptions in BookingApplication

+++ ConsoleBookingSystem/HostService.cs
@@ -1,3 +1,5 @@
+using System.Data;
+
 namespace ConsoleBookingSystem;
 
 public class HostService : IHostService
@@ -37,9 +39,7 @@ public class HostService : IHostService
     public void AddHost(string hostName, string address)
     {
         if (HostExistsByNameAndAddress(hostName, address))
-        {
-            throw new Exception("Host with specified name and address already exists.");
-        }
+            throw new ArgumentException("Host with specified name and address already exists.");
         
         _hostRepository.CreateHost(new CreateHostData { Name = hostName, Address = address });
         
@@ -48,16 +48,15 @@ public class HostService : IHostService
 
     public void RemoveHostById(int hostId)
     {
+        if (hostId <= 0)
+            throw new ArgumentOutOfRangeException(nameof(hostId), "Host ID cannot be zero or negative.");
+        
         if (_hostRepository.FindHostsCount() == 0)
-        {
-            throw new Exception("No Hosts added yet.");
-        }
+            throw new HostNotFoundException();
             
         var host = _hostRepository.FindHostById(hostId);
         if (host == null)
-        {
-            throw new Exception($"Host with ID = {hostId} not found.");
-        }
+            throw new HostNotFoundException("Host with specified ID not found.", hostId);
         
         _hostRepository.RemoveHost(host);
         
@@ -67,10 +66,10 @@ public class HostService : IHostService
     public void EditHostById(int hostId, string updatedHostName, string updatedAddress)
     {
         if (!HostExistsById(hostId))
-            throw new Exception($"Host with ID = {hostId} not found.");
+            throw new HostNotFoundException("Host with specified ID not found.", hostId);
 
         if (HostExistsByNameAndAddress(updatedHostName, updatedAddress, hostId))
-            throw new Exception("Host with specified Name and Address already exists.");
+            throw new ArgumentException("Host with specified Name and Address already exists.");
         
         _hostRepository.UpdateHost(hostId, new UpdateHostData { Name = updatedHostName, Address = updatedAddress });
         
@@ -79,16 +78,8 @@ public class HostService : IHostService
 
     public void SaveChanges()
     {
-        try
-        {
-            _hostRepository.SaveChanges();
-            SetAnyUnsavedChanges(false);
-        }
-        catch (Exception exception)
-        {
-            throw new Exception(exception.Message);
-        }
-        
+        _hostRepository.SaveChanges();
+        SetAnyUnsavedChanges(false);
     }
     
     private void SetAnyUnsavedChanges(bool anyUnsavedChanges)
